refactor some utils
